+++ R/performNormalization.R
@@ -7,14 +7,20 @@
 #' evaluation for log2-fold change, but will alter the original 
 #' enrichment values.
 #' 
-#' @param input.data Enrichment output from \code{\link{escape.matrix}} or
-#' \code{\link{runEscape}}.
-#' @param assay Name of the assay to plot if data is a single-cell object.
+#' @param sc.data Single-cell object or matrix used in the gene set enrichment calculation in 
+#' \code{\link{escape.matrix}} or \code{\link{runEscape}}.
+#' @param enrichment.data The enrichment results from \code{\link{escape.matrix}} 
+#' or \code{\link{runEscape}} (optional)
+#' @param assay Name of the assay to normalize if using a single-cell object
 #' @param gene.sets The gene set library to use to extract 
-#' the individual gene set information from.
+#' the individual gene set information from
 #' @param scale.factor A vector to use for normalizing enrichment scores per cell.
 #' @param make.positive Shift enrichment values to a positive range \strong{TRUE}
 #' for downstream analysis or not \strong{TRUE} (default).
+#' 
+#' @importFrom stringr str_replace_all
+#' @importFrom SeuratObject Assays
+#' @importFrom SummarizedExperiment assays
 #' 
 #' @examples
 #' GS <- list(Bcells = c("MS4A1", "CD79B", "CD79A", "IGH1", "IGH2"),
@@ -32,20 +38,30 @@
 #' 
 #' @return Single-cell object or matrix of normalized enrichment scores
 
-performNormalization <- function(input.data,
-                                 assay = NULL,
+performNormalization <- function(sc.data,
+                                 enrichment.data = NULL,
+                                 assay = "escape",
                                  gene.sets = NULL,
                                  make.positive = FALSE,
                                  scale.factor = NULL) {
-  
-  
-  if(is_seurat_or_se_object(input.data)) {
-    enriched <- .pull.Enrich(input.data, assay)
+  if(!is.null(assay)) {
+    if(is_seurat_object(sc.data)) {
+      assay.present <- assay %in% Assays(sc.data)
+    } else if (is_se_object(sc.data)) {
+      assay.present <- assay %in% assays(sc.data)
+    }
   } else {
-    enriched <- input.data
+    assay.present <- FALSE
   }
   
-  if(!is.null(scale.factor) & length(scale.factor) != dim(input.data)[2]) {
+  
+  if(is_seurat_or_se_object(sc.data) & !is.null(assay) & assay.present) {
+    enriched <- .pull.Enrich(sc.data, assay)
+  } else {
+    enriched <- enrichment.data
+  }
+  
+  if(!is.null(scale.factor) & length(scale.factor) != dim(sc.data)[2]) {
     stop("If using a vector as a scale factor, please ensure the length matches the number of cells.")
   }
   
@@ -55,7 +71,7 @@ performNormalization <- function(input.data,
   egc <- egc[names(egc) %in% colnames(enriched)]
   
   #Isolating the number of genes per cell expressed
-  cnts <- .cntEval(input.data, assay = "RNA", type = "counts")
+  cnts <- .cntEval(sc.data, assay = "RNA", type = "counts")
   
   if(is.null(scale.factor)) {
     print("Calculating features per cell...")
@@ -98,9 +114,12 @@ performNormalization <- function(input.data,
   normalized.enriched <- do.call(cbind, normalized.values)
   colnames(normalized.enriched) <- colnames(enriched)
   
-  if(is_seurat_or_se_object(input.data)) {
-    input.data <- .adding.Enrich(input.data, normalized.enriched, paste0(assay, "_normalized"))
-    return(input.data)
+  if(is_seurat_or_se_object(sc.data)) {
+    if(is.null(assay)) {
+      assay <- "escape"
+    }
+    sc.data <- .adding.Enrich(sc.data, normalized.enriched, paste0(assay, "_normalized"))
+    return(sc.data)
   } else {
     return(normalized.enriched)
   }